+++ Java/src/com/khoilg2008110135/tuan9/dog.java
@@ -7,7 +7,7 @@ public class dog extends Animal{
 	 * @param d
 	 * @param breed
 	 */
-	public dog(String name, double d, String breed) {
+	public dog(String name, int d, String breed) {
 		super(name, d, breed);
 	}
